cv_service/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
torch
torchvision
pillow
cryptography
numpy
